+++ docs/CHANGES.rst
@@ -6,6 +6,8 @@ Change Log
 v19.1.dev0 (unreleased)
 =======================
 
+* Drop support for Python 2.7; minimum version is now 3.4.
+
 * Change to use `black`_ for code style management.
 
   .. _black: https://black.readthedocs.io/en/stable/